Update the simulator to be more accurate
The issue is that for a particular fragment, only CtT or GtA damage can
happen, not both. This is due to how the strand is reconstructed during
the sequencing process.

To fix this then, I gave each fragment a type, either 5' or 3'. This
type indicates which side of the read the "virtual" overhang is allowed
on. This type is determined on fragment creation, with equal weights for
both